requirements.txt
streamlit
requests
beautifulsoup4
duckduckgo-search
python-docx
PyPDF2
pandas
